Implemented Phase 1 of self-scheduling
Now when a BoF hits the interest threshold, an email is sent to the
propser with a token. This token can be used to log in as a 'self
scheduler', who is allowed to schedule only the one session that the
email was about. The token is 3 words out of token-words.txt, currenly a
subset of the Jargon File.

Phase 2 will be a daemon to send emails to the schedulers if someone
doesn't self-schedule a session in X amount of time, so things don't
fall through the cracks.