Implemented signals between bof-o-matic and the daemons to expedite sending
Now SIGUSR1 is sent to the SMS and SMTP daemons when something is
queued, and they'll pick that up within a few seconds and run the queue
immediately, rather than waiting for the next poll. Also implemented a
fix for some race conditions that could crash bom-smsd if it tried to
run the queue at the same time as bom-maild.